decomposed reducers
