+++ src/components/Todo/Todo.jsx
@@ -20,7 +20,7 @@ const Todo = (props) => (
 
 function mapStateToProps(state) {
     return {
-        todos: state.todos
+        todos: state.todo.todos
     }
 }
