+++ src/redux/reducers/todo.js
@@ -1,12 +1,7 @@
 const initialState = {
     todos: [],
-    user: {
-        email: null,
-        password: null,
-    },
 }
-
-export const rootReducer = (state = initialState, action) => {
+export const todo = (state = initialState, action) => {
 
     switch (action.type) {
         case 'MARK_TASK_AS_COMPLETE': {
@@ -41,14 +36,6 @@ export const rootReducer = (state = initialState, action) => {
                 todos: [ ...state.todos, action.payload],
             }
 
-        case 'SET_USER_DATA':
-            const newUserData = {...state.user};
-            newUserData.user = action.payload.email;
-            newUserData.password = action.payload.password;
-            return {
-                ...state,
-                user: newUserData,
-            }
 
         default: return state;
     }
